Enum gurobi::Where

pub enum Where {
    Polling,
    PreSolve {
        coldel: i32,
        rowdel: i32,
        senchg: i32,
        bndchg: i32,
        coecfg: i32,
    },
    Simplex {
        itrcnt: f64,
        objval: f64,
        priminf: f64,
        dualinf: f64,
        ispert: i32,
    },
    MIP {
        objbst: f64,
        objbnd: f64,
        nodcnt: f64,
        solcnt: f64,
        cutcnt: i32,
        nodleft: f64,
        itrcnt: f64,
    },
    MIPSol {
        obj: f64,
        objbst: f64,
        objbnd: f64,
        nodcnt: f64,
        solcnt: f64,
    },
    MIPNode {
        status: i32,
        objbst: f64,
        objbnd: f64,
        nodcnt: f64,
        solcnt: i32,
    },
    Message(String),
    Barrier {
        itrcnt: i32,
        primobj: f64,
        dualobj: f64,
        priminf: f64,
        dualinf: f64,
        compl: f64,
    },
}

Variants
Polling

b

PreSolve

a

Fields

coldel: i32

The number of columns removed by presolve to this point.

rowdel: i32

The number of rows removed by presolve to this point.

senchg: i32

The number of constraint senses changed by presolve to this point.

bndchg: i32

The number of variable bounds changed by presolve to this point.

coecfg: i32

The number of coefficients changed by presolve to this point.

Simplex

c

Fields

itrcnt: f64

Current simplex iteration count.

objval: f64

Current simplex objective value.

priminf: f64

Current primal infeasibility.

dualinf: f64

Current dual infeasibility.

ispert: i32

Is problem current perturbed?

MIP

d

Fields

objbst: f64

Current best objective.

objbnd: f64

Current best objective bound.

nodcnt: f64

Current explored node count.

solcnt: f64

Current count of feasible solutions found.

cutcnt: i32

Current count of cutting planes applied.

nodleft: f64

Current unexplored node count.

itrcnt: f64

Current simplex iteration count.

MIPSol

e

Fields

obj: f64

Objective value for new solution.

objbst: f64

Current best objective.

objbnd: f64

Current best objective bound.

nodcnt: f64

Current explored node count.

solcnt: f64

Current count of feasible solutions found.

MIPNode

f

Fields

status: i32

Optimization status of current MIP node (see the Status Code section for further information).

objbst: f64

Current best objective.

objbnd: f64

Current best objective bound.

nodcnt: f64

Current explored node count.

solcnt: i32

Current count of feasible solutions found.

Message(String)

f

Barrier

d

Fields

itrcnt: i32

Current barrier iteration count.

primobj: f64

Primal objective value for current barrier iterate.

dualobj: f64

Dual objective value for current barrier iterate.

priminf: f64

Primal infeasibility for current barrier iterate.

dualinf: f64

Dual infeasibility for current barrier iterate.

compl: f64

Complementarity violation for current barrier iterate.