Crate gurobi

A crate which provides low-level Rust API of Gurobi Optimizer.

This crate provides wrappers of the Gurobi solver which supports some types of mathematical programming problems (e.g. Linear programming; LP, Mixed Integer Linear Programming; MILP, and so on).

Installation

Before using this crate, you should install Gurobi and obtain a license. The instruction can be found here.

Examples

Work in progress...

Reexports

pub use error::{Error, Result};

pub use env::{param, Env};

pub use model::{attr, Model, Var, LinExpr, QuadExpr};

pub use model::VarType::*;

pub use model::ConstrSense::*;

pub use model::ModelSense::*;

pub use model::SOSType::*;

Modules

env

error

model