Struct gurobi::model::Model

pub struct Model<'a> {
    // some fields omitted
}

Gurobi Model

Methods

impl<'a> Model<'a>

fn new(env: &'a Env, model: *mut GRBmodel) -> Model<'a>

create an empty model which associated with certain environment.

fn update(&mut self) -> Result<()>

apply all modification of the model to process.

fn copy(&self) -> Result<Model>

create a copy of the model

fn optimize(&mut self) -> Result<()>

optimize the model.

fn write(&self, filename: &str) -> Result<()>

write information of the model to file.

fn add_vars<S: Shape>(&mut self, name: &str, vtype: VarType, shape: S) -> Result<Var<S>>

add a set of decision variables to the model.

fn add_constrs<S: Shape>(&mut self, name: &str, expr: LinExpr<S>, sense: ConstrSense, rhs: f64) -> Result<Constr<S>>

add a linear constraint to the model.

fn add_qconstrs<S: Shape>(&mut self, name: &str, expr: QuadExpr<S>, sense: ConstrSense, rhs: f64) -> Result<QConstr<S>>

add a quadratic constraint to the model.

fn set_objective<Expr: Into<QuadExpr<()>>>(&mut self, expr: Expr, sense: ModelSense) -> Result<()>

Set the objective function of the model.

fn get<A: Attr>(&self, attr: A) -> Result<A::Out>

fn set<A: Attr>(&mut self, attr: A, value: A::Out) -> Result<()>

fn get_values<A: AttrArray + Copy, S: Shape, E: Tensor<i32, S>>(&self, attr: A, e: &[&E]) -> Result<Vec<TensorVal<A::Out, S>>>

fn set_values<A: AttrArray + Copy, S: Shape, E: Tensor<i32, S>>(&mut self, attr: A, e: &[&E], val: &[&Tensor<A::Out, S>]) -> Result<()>

fn calc_value<S: Shape, E: Clone + Into<QuadExpr<S>>>(&self, expr: &E) -> Result<Vec<f64>>